+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fix
 - variables in gql Query body are replaced (AnWeber/vscode-httpyac#303)
 - allow defaultHeaders to overwrite accept or user-agent header
+- update label of testItem on httpRegion name change (#315)
 
 
 ## [6.14.0] ( 2024-06-01)